+++ tests/nonmem/test_modelfit_results.py
@@ -15,9 +15,8 @@ def test_ofv(pheno):
     assert res.ofv == 586.27605628188053
 
 
-def test_aic_bic(testdata):
+def test_bic(testdata):
     model = Model.create_model(testdata / 'nonmem' / 'pheno.mod')
-    assert model.modelfit_results.aic == 740.8947268137307
     assert model.modelfit_results.bic == 756.111852398327