+++ CHANGELOG.rst
@@ -1,3 +1,8 @@
+next version
+------------
+
+* Add modeling.calculate_aic
+
 0.45.0 (2022-01-21)
 -------------------